Tests/LibWeb/Text/expected/Editing/textarea-home-end-navigation.txt

--- Home from middle of each line ---
Home: 5 -> 0
Home: 16 -> 11
Home: 25 -> 23
--- End from middle of each line ---
End: 5 -> 10
End: 16 -> 22
End: 25 -> 33
